pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "octopus"
version = "0.1.0"
description = "EmailOctopus Campaign Analytics & Reporting Dashboard"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "EmpowerSaves", email = "dev@empowersaves.com"}
]
keywords = ["emailoctopus", "campaign", "analytics", "reporting", "dashboard"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Flask",
]

dependencies = [
    "Flask>=3.0.0",
    "Werkzeug>=3.0.1",
    "Flask-SQLAlchemy>=3.1.1",
    "Flask-Login>=0.6.3",
    "Flask-WTF>=1.2.1",
    "email-validator>=2.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.26.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]
mongodb = [
    "pymongo>=4.6.1",
    "mongoengine>=0.27.0",
]
reporting = [
    "plotly>=5.18.0",
    "ReportLab>=4.0.7",
    "APScheduler>=3.10.4",
]

[project.scripts]
octopus = "app.cli:main"
octopus-create-user = "scripts.create_user:main"

[project.urls]
Homepage = "https://github.com/empowersaves/octopus"
Documentation = "https://github.com/empowersaves/octopus/blob/main/README.md"
Repository = "https://github.com/empowersaves/octopus"
"Bug Tracker" = "https://github.com/empowersaves/octopus/issues"

[tool.setuptools]
packages = ["app", "src", "scripts"]
include-package-data = true

[tool.setuptools.package-data]
app = ["templates/**/*", "static/**/*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
